up: fixed some cases with star but not all, moved files
