+++ src/tests/test_cmp.c
