+++ src/tests/test_lexer.c
